feat(react): add use case visualizer
add interfaces for all the node types and edge types in a graph
and an adapter which transforms the graph components fetched by
the use case designer.

add ELK layout which determines how the nodes/edges should be
laid out on the UI and supplies them to ReactFlow for final rendering.

Signed-off-by: Satya Krishna Pindiproli <satyak@qti.qualcomm.com>